Clean up file structure for `error-stack`
