+++ packages/libs/error-stack/src/ext/result.rs
